requirements.txt
keboola.component>=1.6.10
keboola.utils
keboola.http-client
pydantic
langchain
langchain-core
langchain-community
langchain-openai
langchain-postgres
langchain-cohere
langchain-aws
langchain-google-vertexai
langchain-huggingface
langchain_pinecone
langchain_qdrant
langchain_redis
langchain_milvus

# Test dependencies
pytest
pytest-asyncio
pytest-mock
freezegun

boto3
cohere
tenacity
pinecone
pymilvus
psycopg-binary
psycopg-pool
#psycopg2-binary
greenlet
asyncpg
qdrant-client
redis
opensearch-py
pgvector
Google-api-core
